+++ exporter/opentelemetry-exporter-otlp-proto-http/tests/test_proto_span_exporter.py
@@ -66,6 +66,7 @@ BASIC_SPAN = _Span(
             "trace_state": {"a": "b", "c": "d"},
             "span_id": 10217189687419569865,
             "trace_id": 67545097771067222548457157018666467027,
+            "trace_flags": 0x00,
         }
     ),
 )